re-mod the draggle item to a new re-usable component :)
