+++ src/components/features/Questions/GraphingCalculator.tsx
@@ -1,12 +1,9 @@
 'use client';
 
 import React, { useEffect, useRef, useState } from 'react';
-import { X } from 'lucide-react';
 import { useCalcOptionModalStore } from '@/store/modals';
 import { DesmosCalculator } from '@/types/desmos';
-
-// Define interface for Desmos Calculator instance
-
+import DraggableItem from '@/components/DraggableItem';
 
 // Define the Desmos type
 declare global {
@@ -17,22 +14,11 @@ declare global {
   }
 }
 
-interface Position {
-  x: number;
-  y: number;
-}
 
-interface GraphCalculatorProps {
-  handleEndState: () => void;
-}
-
-const GraphCalculator: React.FC<GraphCalculatorProps> = ({ handleEndState }) => {
+const GraphCalculator = () => {
   const closeModal = useCalcOptionModalStore((state) => state.closeModal);
   const containerRef = useRef<HTMLDivElement>(null);
   const [isReady, setIsReady] = useState(false);
-  const [position, setPosition] = useState<Position>({ x: 50, y: 50 });
-  const [isDragging, setIsDragging] = useState(false);
-  const [dragOffset, setDragOffset] = useState<Position>({ x: 0, y: 0 });
   const calculatorRef = useRef<DesmosCalculator | null>(null);
 
   useEffect(() => {
@@ -62,69 +48,20 @@ const GraphCalculator: React.FC<GraphCalculatorProps> = ({ handleEndState }) => 
     };
   }, []);
 
-  const handleMouseDown = (e: React.MouseEvent<HTMLDivElement>) => {
-    if (e.target instanceof HTMLElement && e.target.closest('.calculator-content')) return;
-    setIsDragging(true);
-    setDragOffset({
-      x: e.clientX - position.x,
-      y: e.clientY - position.y,
-    });
-  };
 
-  useEffect(() => {
-    const handleMouseMove = (e: MouseEvent) => {
-      if (!isDragging) return;
-      setPosition({
-        x: e.clientX - dragOffset.x,
-        y: e.clientY - dragOffset.y,
-      });
-    };
-
-    const handleMouseUp = () => {
-      setIsDragging(false);
-    };
-
-    if (isDragging) {
-      window.addEventListener('mousemove', handleMouseMove);
-      window.addEventListener('mouseup', handleMouseUp);
-    }
-
-    return () => {
-      window.removeEventListener('mousemove', handleMouseMove);
-      window.removeEventListener('mouseup', handleMouseUp);
-    };
-  }, [isDragging, dragOffset]);
 
   return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center">
-      <div
-        className="bg-white rounded-lg shadow-xl"
-        style={{
-          position: 'absolute',
-          left: `${position.x}px`,
-          top: `${position.y}px`,
-        }}
-      >
-        <div
-          className="flex items-center justify-between p-2 bg-gray-100 cursor-move"
-          onMouseDown={handleMouseDown}
-        >
-          <h3 className="font-medium">Graphing Calculator</h3>
-          <button
-            onClick={handleEndState}
-            className="p-1 hover:bg-gray-200 rounded transition-colors"
-          >
-            <X size={20} />
-          </button>
-        </div>
+    <DraggableItem 
+      title='Graphing Calculator'
+      content={
         <div className="calculator-content">
           <div
             ref={containerRef}
             className={`w-[600px] h-[400px] ${!isReady ? 'invisible' : ''}`}
           />
         </div>
-      </div>
-    </div>
+      }
+    />
   );
 };
