+++ src/store/modals.ts
@@ -1,9 +1,15 @@
+import { calc } from '@/types/calc';
 import { create } from 'zustand'
 
 interface ModalProps {
     isOpen: boolean
     openModal: () => void;
     closeModal: () => void;
+}
+
+interface CalcMode {
+  mode: calc;
+  setMode: (mode: calc) => void;
 }
 
 export const useScoreModalStore = create<ModalProps>((set) => ({
@@ -29,3 +35,8 @@ export const useCalcOptionModalStore = create<ModalProps>((set) => ({
   openModal: () => set(() => ({ isOpen: true })),
   closeModal: () => set(() => ({ isOpen: false }))
 }))
+
+export const useCalcModeModalStore = create<CalcMode>((set) => ({
+  mode: "none",
+  setMode: (mode: calc) => set(() => ({ mode }))
+}));
